requirements.txt
beautifulsoup4
fastapi
jinja2
lxml
playwright
pywebview
python-multipart
requests
sqlalchemy
uvicorn
